requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.5
asyncio>=3.4.3
openai>=1.3.0
requests>=2.28.0
